requirements.txt
Flask==3.0.3
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
gunicorn==21.2.0
defusedxml==0.7.1
python-dotenv==1.0.1